+++ apps/meteor/app/integrations/server/api/api.ts
@@ -14,7 +14,7 @@ import { isPlainObject } from '../../../../lib/utils/isPlainObject';
 import { APIClass } from '../../../api/server/ApiClass';
 import type { RateLimiterOptions } from '../../../api/server/api';
 import { API, defaultRateLimiterOptions } from '../../../api/server/api';
-import type { FailureResult, PartialThis, SuccessResult, UnavailableResult } from '../../../api/server/definition';
+import type { FailureResult, GenericRouteExecutionContext, SuccessResult, UnavailableResult } from '../../../api/server/definition';
 import type { WebhookResponseItem } from '../../../lib/server/functions/processWebhookMessage';
 import { processWebhookMessage } from '../../../lib/server/functions/processWebhookMessage';
 import { settings } from '../../../settings/server';
@@ -41,11 +41,10 @@ type IntegrationOptions = {
 	};
 };
 
-type IntegrationThis = Omit<PartialThis, 'user'> & {
+type IntegrationThis = GenericRouteExecutionContext & {
 	request: Request & {
 		integration: IIncomingIntegration;
 	};
-	urlParams: Record<string, string>;
 	user: IUser & { username: RequiredField<IUser, 'username'> };
 };
 
@@ -154,21 +153,22 @@ async function executeIntegrationRest(
 		const contentRaw = Buffer.concat(buffers).toString('utf8');
 		const protocol = `${this.request.headers.get('x-forwarded-proto')}:` || 'http:';
 		const url = new URL(this.request.url, `${protocol}//${this.request.headers.get('host')}`);
+		const query = isPlainObject(this.queryParams) ? this.queryParams : {};
 
 		const request = {
 			url: {
+				query,
 				hash: url.hash,
 				search: url.search,
-				query: this.queryParams,
 				pathname: url.pathname,
 				path: this.request.url,
 			},
 			url_raw: this.request.url,
 			url_params: this.urlParams,
-			content: this.bodyParams,
+			content: bodyParams,
 			content_raw: contentRaw,
 			headers: Object.fromEntries(this.request.headers.entries()),
-			body: this.bodyParams,
+			body: bodyParams,
 			user: {
 				_id: this.user._id,
 				name: this.user.name || '',
@@ -318,8 +318,8 @@ function integrationInfoRest(): { statusCode: number; body: { success: boolean }
 }
 
 class WebHookAPI extends APIClass<'/hooks'> {
-	override async authenticatedRoute(this: IntegrationThis): Promise<IUser | null> {
-		const { integrationId, token } = this.urlParams;
+	override async authenticatedRoute(routeContext: IntegrationThis): Promise<IUser | null> {
+		const { integrationId, token } = routeContext.urlParams;
 		const integration = await Integrations.findOneByIdAndToken<IIncomingIntegration>(integrationId, decodeURIComponent(token));
 
 		if (!integration) {
@@ -328,9 +328,9 @@ class WebHookAPI extends APIClass<'/hooks'> {
 			throw new Error('Invalid integration id or token provided.');
 		}
 
-		this.request.integration = integration;
+		routeContext.request.integration = integration;
 
-		return Users.findOneById(this.request.integration.userId);
+		return Users.findOneById(routeContext.request.integration.userId);
 	}
 
 	override shouldAddRateLimitToRoute(options: { rateLimiterOptions?: RateLimiterOptions | boolean }): boolean {