+++ apps/meteor/client/views/room/Header/Header.tsx
@@ -1,38 +1,24 @@
+import { isInviteSubscription } from '@rocket.chat/core-typings';
 import type { IRoom, ISubscription } from '@rocket.chat/core-typings';
-import { isDirectMessageRoom, isVoipRoom, isInviteSubscription } from '@rocket.chat/core-typings';
-import { HeaderToolbar, SidebarToggler } from '@rocket.chat/ui-client';
 import { useLayout, useSetting } from '@rocket.chat/ui-contexts';
 import type { ReactElement } from 'react';
-import { lazy, memo, useMemo } from 'react';
+import { lazy, memo } from 'react';
 
 const RoomInviteHeader = lazy(() => import('./RoomInviteHeader'));
 const OmnichannelRoomHeader = lazy(() => import('./Omnichannel/OmnichannelRoomHeader'));
-const VoipRoomHeader = lazy(() => import('./Omnichannel/VoipRoomHeader'));
 const RoomHeaderE2EESetup = lazy(() => import('./RoomHeaderE2EESetup'));
-const DirectRoomHeader = lazy(() => import('./DirectRoomHeader'));
 const RoomHeader = lazy(() => import('./RoomHeader'));
 
-type HeaderProps<T> = {
-	room: T;
+type HeaderProps = {
+	room: IRoom;
 	subscription?: ISubscription;
 };
 
-const Header = ({ room, subscription }: HeaderProps<IRoom>): ReactElement | null => {
-	const { isMobile, isEmbedded, showTopNavbarEmbeddedLayout } = useLayout();
+const Header = ({ room, subscription }: HeaderProps): ReactElement | null => {
+	const { isEmbedded, showTopNavbarEmbeddedLayout } = useLayout();
 	const encrypted = Boolean(room.encrypted);
 	const unencryptedMessagesAllowed = useSetting('E2E_Allow_Unencrypted_Messages', false);
 	const shouldDisplayE2EESetup = encrypted && !unencryptedMessagesAllowed;
-
-	const slots = useMemo(
-		() => ({
-			start: isMobile && (
-				<HeaderToolbar>
-					<SidebarToggler />
-				</HeaderToolbar>
-			),
-		}),
-		[isMobile],
-	);
 
 	if (isEmbedded && !showTopNavbarEmbeddedLayout) {
 		return null;
@@ -43,22 +29,14 @@ const Header = ({ room, subscription }: HeaderProps<IRoom>): ReactElement | null
 	}
 
 	if (room.t === 'l') {
-		return <OmnichannelRoomHeader slots={slots} />;
-	}
-
-	if (isVoipRoom(room)) {
-		return <VoipRoomHeader slots={slots} room={room} />;
+		return <OmnichannelRoomHeader />;
 	}
 
 	if (shouldDisplayE2EESetup) {
-		return <RoomHeaderE2EESetup room={room} topic={room.topic} slots={slots} />;
+		return <RoomHeaderE2EESetup room={room} />;
 	}
 
-	if (isDirectMessageRoom(room) && (room.uids?.length ?? 0) < 3) {
-		return <DirectRoomHeader slots={slots} room={room} />;
-	}
-
-	return <RoomHeader room={room} topic={room.topic} slots={slots} />;
+	return <RoomHeader room={room} />;
 };
 
 export default memo(Header);