+++ apps/meteor/tests/e2e/omnichannel/omnichannel-close-chat.spec.ts
@@ -42,7 +42,7 @@ test.describe('Omnichannel close chat', () => {
 		});
 
 		await test.step('Expect to have 1 omnichannel assigned to agent 1', async () => {
-			await agent.poHomeOmnichannel.sidenav.openChat(newVisitor.name);
+			await agent.poHomeOmnichannel.navbar.openChat(newVisitor.name);
 		});
 
 		await test.step('Expect to be able to close an omnichannel to conversation', async () => {