+++ packages/models/src/models/Users.ts
@@ -92,7 +92,6 @@ export class UsersRaw extends BaseRaw<IUser, DefaultFields<IUser>> implements IU
 			{ key: { 'services.saml.inResponseTo': 1 } },
 			{ key: { openBusinessHours: 1 }, sparse: true },
 			{ key: { statusLivechat: 1 }, sparse: true },
-			{ key: { extension: 1 }, sparse: true, unique: true },
 			{ key: { freeSwitchExtension: 1 }, sparse: true, unique: true },
 			{ key: { language: 1 }, sparse: true },
 			{ key: { 'active': 1, 'services.email2fa.enabled': 1 }, sparse: true }, // used by statistics
@@ -1476,78 +1475,6 @@ export class UsersRaw extends BaseRaw<IUser, DefaultFields<IUser>> implements IU
 		return this.updateOne(query, update);
 	}
 
-	// Voip functions
-	findOneByAgentUsername(username: string, options?: FindOptions<IUser>) {
-		const query = { username, roles: 'livechat-agent' };
-
-		return this.findOne(query, options);
-	}
-
-	findOneByExtension(extension: string, options?: FindOptions<IUser>) {
-		const query = {
-			extension,
-		};
-
-		return this.findOne(query, options);
-	}
-
-	findByExtensions(extensions: string[], options?: FindOptions<IUser>) {
-		const query = {
-			extension: {
-				$in: extensions,
-			},
-		};
-
-		return this.find(query, options);
-	}
-
-	getVoipExtensionByUserId(userId: IUser['_id'], options?: FindOptions<IUser>) {
-		const query = {
-			_id: userId,
-			extension: { $exists: true },
-		};
-		return this.findOne(query, options);
-	}
-
-	setExtension(userId: IUser['_id'], extension: string) {
-		const query = {
-			_id: userId,
-		};
-
-		const update = {
-			$set: {
-				extension,
-			},
-		};
-		return this.updateOne(query, update);
-	}
-
-	unsetExtension(userId: IUser['_id']) {
-		const query = {
-			_id: userId,
-		};
-		const update = {
-			$unset: {
-				extension: 1 as const,
-			},
-		};
-		return this.updateOne(query, update);
-	}
-
-	getAvailableAgentsIncludingExt<T extends Document = ILivechatAgent>(includeExt?: string, text?: string, options?: FindOptions<IUser>) {
-		const query = {
-			roles: { $in: ['livechat-agent'] },
-			$and: [
-				...(text?.trim()
-					? [{ $or: [{ username: new RegExp(escapeRegExp(text), 'i') }, { name: new RegExp(escapeRegExp(text), 'i') }] }]
-					: []),
-				{ $or: [{ extension: { $exists: false } }, ...(includeExt ? [{ extension: includeExt }] : [])] },
-			],
-		};
-
-		return this.findPaginated<T>(query, options);
-	}
-
 	findActiveUsersTOTPEnable(options?: FindOptions<IUser>) {
 		const query = {
 			'active': true,
@@ -2068,7 +1995,7 @@ export class UsersRaw extends BaseRaw<IUser, DefaultFields<IUser>> implements IU
 		return this.updateOne(query, update);
 	}
 
-	makeAgentUnavailableAndUnsetExtension(userId: IUser['_id']) {
+	makeAgentUnavailable(userId: IUser['_id']) {
 		const query = {
 			_id: userId,
 			roles: 'livechat-agent',
@@ -2077,9 +2004,6 @@ export class UsersRaw extends BaseRaw<IUser, DefaultFields<IUser>> implements IU
 		const update = {
 			$set: {
 				statusLivechat: ILivechatAgentStatus.NOT_AVAILABLE,
-			},
-			$unset: {
-				extension: 1 as const,
 			},
 		};
 
@@ -2804,34 +2728,6 @@ export class UsersRaw extends BaseRaw<IUser, DefaultFields<IUser>> implements IU
 		);
 	}
 
-	findOneByFreeSwitchExtensions<T extends Document = IUser>(freeSwitchExtensions: string[], options: FindOptions<IUser> = {}) {
-		return this.findOne<T>(
-			{
-				freeSwitchExtension: { $in: freeSwitchExtensions },
-			},
-			options,
-		);
-	}
-
-	findAssignedFreeSwitchExtensions() {
-		return this.findUsersWithAssignedFreeSwitchExtensions({
-			projection: {
-				freeSwitchExtension: 1,
-			},
-		}).map(({ freeSwitchExtension }) => freeSwitchExtension);
-	}
-
-	findUsersWithAssignedFreeSwitchExtensions<T extends Document = IUser>(options: FindOptions<IUser> = {}) {
-		return this.find<T>(
-			{
-				freeSwitchExtension: {
-					$exists: true,
-				},
-			},
-			options,
-		);
-	}
-
 	// UPDATE
 	addImportIds(_id: IUser['_id'], importIds: string[]) {
 		importIds = ([] as string[]).concat(importIds);
@@ -3256,17 +3152,6 @@ export class UsersRaw extends BaseRaw<IUser, DefaultFields<IUser>> implements IU
 		);
 	}
 
-	async setFreeSwitchExtension(_id: IUser['_id'], extension?: string) {
-		return this.updateOne(
-			{
-				_id,
-			},
-			{
-				...(extension ? { $set: { freeSwitchExtension: extension } } : { $unset: { freeSwitchExtension: 1 } }),
-			},
-		);
-	}
-
 	// INSERT
 	create(data: InsertionModel<IUser>) {
 		const user = {
@@ -3424,7 +3309,6 @@ export class UsersRaw extends BaseRaw<IUser, DefaultFields<IUser>> implements IU
 			$unset: {
 				livechat: 1,
 				statusLivechat: 1,
-				extension: 1,
 				openBusinessHours: 1,
 			},
 		};