+++ packages/ui-client/src/components/Page/PageHeaderNoShadow.tsx
@@ -4,7 +4,6 @@ import type { ComponentPropsWithoutRef, ReactNode } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { useDocumentTitle } from '../../hooks';
-import { FeaturePreview, FeaturePreviewOff, FeaturePreviewOn } from '../FeaturePreview';
 import { HeaderToolbar } from '../Header';
 import { SidebarToggler } from '../SidebarToggler';
 
@@ -32,22 +31,11 @@ const PageHeaderNoShadow = ({ children = undefined, title, onClickBack, ...props
 				alignItems='center'
 				color='default'
 			>
-				<FeaturePreview feature='newNavigation'>
-					<FeaturePreviewOff>
-						{sidebar.shouldToggle ? (
-							<HeaderToolbar>
-								<SidebarToggler />
-							</HeaderToolbar>
-						) : null}
-					</FeaturePreviewOff>
-					<FeaturePreviewOn>
-						{sidebar.shouldToggle && isEmbedded ? (
-							<HeaderToolbar>
-								<SidebarToggler />
-							</HeaderToolbar>
-						) : null}
-					</FeaturePreviewOn>
-				</FeaturePreview>
+				{sidebar.shouldToggle && isEmbedded ? (
+					<HeaderToolbar>
+						<SidebarToggler />
+					</HeaderToolbar>
+				) : null}
 				{onClickBack && <IconButton small mie={8} icon='arrow-back' onClick={onClickBack} title={t('Back')} />}
 				<Box is='h1' fontScale='h2' flexGrow={1} data-qa-type='PageHeader-title'>
 					{title}